+++ R/S4classes.R
@@ -48,6 +48,7 @@ NULL
 #'   plenary protocols), or `NA`.
 #' @slot name Full name of the corpus that may be more expressive than
 #'   the corpus ID.
+#' @slot xml Object of class `character`, whether the xml is "flat" or "nested".
 #' @slot encoding The encoding of the corpus, given as a length-one
 #'   `character` vector (usually 'utf8' or 'latin1').
 #' @slot size Number of tokens (size) of the corpus, a length-one `integer`
@@ -125,7 +126,8 @@ setClass(
     type = "character",
     encoding = "character",
     name = "character",
-    size = "integer"
+    size = "integer",
+    xml = "character"
   )
 )
 
@@ -842,7 +844,13 @@ setMethod("name", "corpus", function(x) x@name)
 #' R <- as.regions(P)
 #' @aliases regions-class
 #' @family classes to manage corpora
-setClass("regions", slots = c(cpos = "matrix"), contains = "corpus")
+setClass(
+  "regions",
+  slots = c(
+    cpos = "matrix"
+  ),
+  contains = "corpus"
+)